requirements.txt
pytest==2.6.1
pytest-cov==1.8.1
tox==1.7.2
isort==4.1.1
ipython==3.2.1
wheel==0.24.0
python-coveralls==2.5.0
hug==1.3.0
hot-redis==0.3
gunicorn==19.3.0
Pillow==2.9.0
sh==1.11
Cython==0.22.1
